requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
djangorestframework==3.13.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
python-weather==0.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
yarl==1.8.1
